+++ src/plugins/curator/services/WidgetHasSearchRequestSubscriber.js
@@ -6,7 +6,6 @@ import { getFormMeta } from 'redux-form';
 
 import EventSubscriber from '@gdbots/pbjx/EventSubscriber';
 import getDatePickerFieldError from '@triniti/cms/components/date-picker-field/getDatePickerFieldError';
-import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 import SearchArticlesRequestV1Mixin from '@triniti/schemas/triniti/news/mixin/search-articles-request/SearchArticlesRequestV1Mixin';
 import SearchArticlesSort from '@triniti/schemas/triniti/news/enums/SearchArticlesSort';
 import SearchGalleriesRequestV1Mixin from '@triniti/schemas/triniti/curator/mixin/search-galleries-request/SearchGalleriesRequestV1Mixin';
@@ -95,9 +94,8 @@ export default class WidgetHasSearchRequestSubscriber extends EventSubscriber {
           value: type,
         }));
 
-        const timelineRef = searchRequest.get('timeline_ref') ? searchRequest.get('timeline_ref').toString() : '';
-        if (timelineRef) {
-          data[camelCase(searchRequestType)].timelineRefs = [timelineRef];
+        if (searchRequest.has('timeline_ref')) {
+          data[camelCase(searchRequestType)].timelineRefs = [searchRequest.get('timeline_ref')];
         }
 
         if (searchRequest.has('gallery_ref')) {