+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # CHANGELOG
 
 
+## vN.N.N
+* Add More Fields On Designated Screens (display_title, credit_url, cta_text, cta_url, etc.)
+
+
 ## v0.3.1
 * Add support for livestream static label
-* Add More Fields On Designated Screens (display_title, credit_url, cta_text, cta_url, etc.)
+* Fix unit test error due to unresolved config file.
+* Fix `WidgetHasSearchRequestSubscriber` `timeline_ref` handling.
 
 
 ## v0.3.0
@@ -22,6 +27,7 @@
 * Add Slotting On Teasers Screen
 * Add checkbox to enable or disable SmartNews.
 * Remove livestream screen breadcrumbs
+* Fixed broken artist embed in Spotify Embed Block
 
 
 ## v0.2.16